requirements.txt
Django==1.6.2
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.19
Mezzanine==3.0.9
Pillow==2.3.1
Werkzeug==0.9.4
argparse==1.2.1
bleach==1.4
django-ajax-selects==1.3.3
django-appconf==0.6
django-bootstrap3-datetimepicker==2.2.3
-e git+git@github.com:riklaunim/django-ckeditor.git@c1d42a504a944f22c31ba6256503a45745a3be52#egg=django_ckeditor_updated-master
django-compressor==1.3
django-debug-toolbar==1.0.1
django-inplaceedit==1.3.0
django-inplaceedit-bootstrap==0.1.1
django-inplaceedit-extra-fields==0.5.0
filebrowser-safe==0.3.2
future==0.11.4
grappelli-safe==0.3.7
html5lib==0.999
image==1.2
itsdangerous==0.24
oauthlib==0.6.1
pytz==2014.2
requests==2.2.1
requests-oauthlib==0.4.0
requirements==0.1
six==1.6.1
sorl-thumbnail==11.12
sqlparse==0.1.11
tzlocal==1.1.1
wsgiref==0.1.2
